[MIG] End migration of payment_payzen
index.html was added to help understand what does the module.

One test was writtent but need more in the future regarding how odoo
implements theirs

Redirection after staying to much time to Payzen side generate a 400
but I can't fix this, thought it was possible with a 303 but no.